+++ operator/internal/controller/pulumi/common.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shared
+package pulumi
 
 type ComponentType = string
