+++ operator/internal/controller/pulumi/stack_controller.go
@@ -1325,8 +1325,8 @@ func nameForWorkspace(stack *metav1.ObjectMeta) string {
 // labelsForWorkspace returns the labels for the direct children of the stack (workspaces and updates).
 func labelsForWorkspace(stack *metav1.ObjectMeta) map[string]string {
 	return map[string]string{
-		shared.ComponentAnnotation: shared.StackComponent,
-		shared.StackNameAnnotation: stack.Name,
+		ComponentAnnotation: StackComponent,
+		StackNameAnnotation: stack.Name,
 	}
 }
